requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.2.0
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
docutils==0.20.1
einops==0.6.1
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
gdown==4.7.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.1
keyring==24.2.0
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
onnxruntime==1.15.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pkginfo==1.9.6
protobuf==4.23.4
psutil==5.9.5
pydantic==2.0.2
pydantic_core==2.1.2
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
readme-renderer==40.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.2
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
sympy==1.12
threadpoolctl==3.1.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.65.0
twine==4.0.2
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.131
urllib3==2.0.3
webencodings==0.5.1
wget==3.2
zipp==3.15.0
